requirements.txt
aiosqlite==0.17.0
asyncpg==0.28.0
iso8601==1.1.0
pypika-tortoise==0.1.6
pytz==2023.3.post1
tortoise-orm==0.20.0
typing_extensions==4.7.1
